Removing auto-configuration for pubsub, statestore
