+++ dapr-spring/dapr-spring-boot-autoconfigure/src/main/java/io/dapr/spring/boot/autoconfigure/pubsub/DaprPubSubProperties.java
@@ -15,6 +15,7 @@ package io.dapr.spring.boot.autoconfigure.pubsub;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 
+
 @ConfigurationProperties(prefix = DaprPubSubProperties.CONFIG_PREFIX)
 public class DaprPubSubProperties {
