+++ dapr-spring/dapr-spring-boot-autoconfigure/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,0 +1,1 @@
+io.dapr.spring.boot.autoconfigure.client.DaprClientAutoConfiguration
