--- sdk-tests/src/test/java/io/dapr/it/spring/boot/autoconfigure/AbstractDaprSpringBootAutoconfigureBaseIT.java
@@ -1,52 +1,0 @@
-/*
- * Copyright 2024 The Dapr Authors
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *     http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package io.dapr.it.spring.boot.autoconfigure;
-
-import io.dapr.testcontainers.Component;
-import io.dapr.testcontainers.DaprContainer;
-import io.dapr.testcontainers.DaprLogLevel;
-import org.junit.jupiter.api.BeforeAll;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.testcontainers.containers.Network;
-import org.testcontainers.containers.output.Slf4jLogConsumer;
-import org.testcontainers.junit.jupiter.Container;
-import org.testcontainers.junit.jupiter.Testcontainers;
-
-import java.util.Collections;
-
-@Testcontainers
-public abstract class AbstractDaprSpringBootAutoconfigureBaseIT {
-
-  public static Network DAPR_NETWORK = Network.newNetwork();
-
-  @Container
-  public static DaprContainer DAPR_CONTAINER = new DaprContainer("daprio/daprd:1.13.2")
-      .withAppName("local-dapr-app")
-      .withNetwork(DAPR_NETWORK)
-      .withComponent(new Component("pubsub", "pubsub.in-memory", "v1", Collections.emptyMap()))
-      .withAppPort(8080)
-      .withDaprLogLevel(DaprLogLevel.DEBUG)
-      .withLogConsumer(outputFrame -> System.out.println(outputFrame.getUtf8String()))
-      .withAppChannelAddress("host.testcontainers.internal");
-
-  @BeforeAll
-  static void beforeAll() {
-    org.testcontainers.Testcontainers.exposeHostPorts(8080);
-
-    System.setProperty("dapr.grpc.port", Integer.toString(DAPR_CONTAINER.getGrpcPort()));
-    System.setProperty("dapr.http.port", Integer.toString(DAPR_CONTAINER.getHttpPort()));
-  }
-
-}
