+++ sdk-tests/src/test/java/io/dapr/it/spring/data/DaprSpringDataConstants.java
@@ -1,0 +1,10 @@
+package io.dapr.it.spring.data;
+
+public class DaprSpringDataConstants {
+  private DaprSpringDataConstants() {
+  }
+
+  public static final String STATE_STORE_NAME = "kvstore";
+  public static final String BINDING_NAME = "kvbinding";
+  public static final String PUBSUB_NAME = "pubsub";
+}
